+++ packages/ui/src/ui/pages/query-tracker/QueriesList/QueriesTutorialList/index.tsx
@@ -4,7 +4,7 @@ import {Icon, List, ListItemData, Loader, Text} from '@gravity-ui/uikit';
 import {QueryItem} from '../../module/api';
 import {useQueryList} from '../../hooks/QueriesList';
 import {useQueryNavigation} from '../../hooks/Query';
-import tutorialIcon from '../../../../../../img/svg/learn.svg';
+import tutorialIcon from '../../../../assets/img/svg/learn.svg';
 import './index.scss';
 
 const itemCn = cn('query-tutorial-item');